+++ src/jsonrpccpp/client/connectors/windowstcpsocketclient.cpp
@@ -74,7 +74,7 @@ void WindowsTcpSocketClient::SendRPCMessage(const std::string& message, std::str
 			cerr << message << endl;
 			throw JsonRpcException(Errors::ERROR_CLIENT_CONNECTOR, message);
 		}
-		else if(byteWritten < toSend.size())
+		else if(static_cast<unsigned int>(byteWritten) < toSend.size())
 		{
 			int len = toSend.size() - byteWritten;
 			toSend = toSend.substr(byteWritten + sizeof(char), len);
